requirements.txt
Django
django-ninja
pydantic
django-ninja-jwt[crypto]